# feat: optimize voluntary exit verification with concurrent processing