## Summary
- Implemented parallel file reading and concurrent signature verification for voluntary exit verification
- Achieved significant performance improvements, especially for large batches of exit files
- Added comprehensive benchmarks to measure and validate the optimizations

## Performance Improvements

The optimizations show substantial performance gains that scale with the number of files:

| File Count | Sequential | Optimized | Improvement |
|------------|------------|-----------|-------------|
| 10 files   | 256μs      | 176μs     | 31% faster  |
| 50 files   | 995μs      | 667μs     | 33% faster  |
| 100 files  | 1,967μs    | 893μs     | 55% faster  |
| 500 files  | 10,122μs   | 4,438μs   | 56% faster  |

## Changes

1. **Parallel File Reading**
   - Implemented worker pool pattern using goroutines
   - Number of workers scales with available CPU cores
   - Thread-safe result collection with proper synchronization

2. **Concurrent Signature Verification**
   - Parallel verification of exit signatures using worker pools
   - Maintains correctness while improving throughput
   - Proper error handling and context preservation

3. **Benchmarks**
   - Added comprehensive benchmarks comparing sequential vs optimized implementations
   - Tests with varying file counts (10, 50, 100, 500) to demonstrate scaling benefits

## Test plan
- [x] All existing tests pass
- [x] Added benchmarks to measure performance improvements
- [x] Verified memory usage remains comparable (no significant overhead)
- [x] Tested with various file counts to ensure correctness

🤖 Generated with [Claude Code](https://claude.ai/code)